microapi/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "micromonte"
authors = ["nika110 <nika.asatiani.1@btu.edu.ge>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.1"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
psycopg2-binary = "^2.9.6"
pydantic = "^1.10.7"
tortoise-orm = "^0.19.3"
aerich = "^0.7.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
python-multipart = "^0.0.6"
pytest-asyncio = "^0.21.0"
faker = "^18.4.0"
asynctest = "^0.13.0"
boto3 = "^1.26.117"
httpx = "^0.24.0"
asyncpg = "^0.27.0"
grpcio-tools = "^1.54.0"
grpcio = "^1.54.0"
redis = {extras = ["hiredis"], version = "^4.5.4"}
aiohttp = "^3.8.4"



[tool.aerich]
tortoise_orm = "core.config.settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

